#4 Adding index controller test
