+++ src/main/java/com/josdem/gmail/controller/IndexController.java
@@ -27,50 +27,50 @@ import org.springframework.web.servlet.ModelAndView;
 @Controller
 public class IndexController {
 
-    @Value("${token}")
-    private String token;
+  @Value("${token}")
+  private String token;
 
-    @GetMapping("/")
-    public String index() {
-        log.info("Calling index");
-        return "index";
-    }
+  @GetMapping("/")
+  public String index() {
+    log.info("Calling index");
+    return "index";
+  }
 
-    @GetMapping("/command-line")
-    public String commandLine() {
-        log.info("Calling command line");
-        return "command_line";
-    }
+  @GetMapping("/command-line")
+  public String commandLine() {
+    log.info("Calling command line");
+    return "command_line";
+  }
 
-    @GetMapping("/form")
-    public ModelAndView form() {
-        log.info("Calling form");
-        var modelAndView = new ModelAndView("form");
-        var message = new MessageCommand();
-        message.setToken(token);
-        modelAndView.addObject("message", message);
-        return modelAndView;
-    }
+  @GetMapping("/form")
+  public ModelAndView form() {
+    log.info("Calling form");
+    var modelAndView = new ModelAndView("form");
+    var message = new MessageCommand();
+    message.setToken(token);
+    modelAndView.addObject("message", message);
+    return modelAndView;
+  }
 
-    @GetMapping("/batch")
-    public String batch() {
-        log.info("Calling batch");
-        return "batch";
-    }
+  @GetMapping("/batch")
+  public String batch() {
+    log.info("Calling batch");
+    return "batch";
+  }
 
-    @GetMapping("/contact")
-    public ModelAndView contact() {
-        log.info("Calling contact");
-        var modelAndView = new ModelAndView("contact");
-        var message = new MessageCommand();
-        message.setToken(token);
-        modelAndView.addObject("message", message);
-        return modelAndView;
-    }
+  @GetMapping("/contact")
+  public ModelAndView contact() {
+    log.info("Calling contact");
+    var modelAndView = new ModelAndView("contact");
+    var message = new MessageCommand();
+    message.setToken(token);
+    modelAndView.addObject("message", message);
+    return modelAndView;
+  }
 
-    @GetMapping("/flyer")
-    public String flyer() {
-        log.info("Calling flyer");
-        return "flyer";
-    }
+  @GetMapping("/flyer")
+  public String flyer() {
+    log.info("Calling flyer");
+    return "flyer";
+  }
 }
